backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
numpy==2.4.1
pandas==3.0.0
pillow==12.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
